+++ test/Validator/LanguageTest.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace TxTextControlTest\ReportingCloud\Validator;
+
+use PHPUnit_Framework_TestCase;
+use TxTextControl\ReportingCloud\Validator\Language as Validator;
+
+class LanguageTest extends PHPUnit_Framework_TestCase
+{
+    protected $validator;
+
+    public function setUp()
+    {
+        $this->validator = new Validator();
+    }
+
+    public function testValid()
+    {
+        $this->assertTrue($this->validator->isValid('de_CH_frami.dic'));
+        $this->assertTrue($this->validator->isValid('pt_BR.dic'));
+        $this->assertTrue($this->validator->isValid('nb_NO.dic'));
+    }
+
+    public function testInvalid()
+    {
+        $this->assertFalse($this->validator->isValid('pt_BR'));
+        $this->assertArrayHasKey(Validator::INVALID_LANGUAGE, $this->validator->getMessages());
+
+        $this->assertFalse($this->validator->isValid('pt'));
+        $this->assertArrayHasKey(Validator::INVALID_LANGUAGE, $this->validator->getMessages());
+
+        $this->assertFalse($this->validator->isValid('0'));
+        $this->assertArrayHasKey(Validator::INVALID_LANGUAGE, $this->validator->getMessages());
+    }
+
+    public function testConstructor()
+    {
+        $validator = new Validator(null);
+
+        $this->assertTrue($validator->isValid('en_US.dic'));
+    }
+}