+++ test/Validator/TypeArrayTest.php
@@ -34,18 +34,18 @@ class TypeArrayTest extends PHPUnit_Framework_TestCase
     public function testInvalidType()
     {
         $this->assertFalse($this->validator->isValid('0'));
-        $this->assertArrayHasKey('invalidType', $this->validator->getMessages());
+        $this->assertArrayHasKey(Validator::INVALID_TYPE, $this->validator->getMessages());
 
         $this->assertFalse($this->validator->isValid(0));
-        $this->assertArrayHasKey('invalidType', $this->validator->getMessages());
+        $this->assertArrayHasKey(Validator::INVALID_TYPE, $this->validator->getMessages());
 
         $this->assertFalse($this->validator->isValid(1));
-        $this->assertArrayHasKey('invalidType', $this->validator->getMessages());
+        $this->assertArrayHasKey(Validator::INVALID_TYPE, $this->validator->getMessages());
 
         $this->assertFalse($this->validator->isValid(false));
-        $this->assertArrayHasKey('invalidType', $this->validator->getMessages());
+        $this->assertArrayHasKey(Validator::INVALID_TYPE, $this->validator->getMessages());
 
         $this->assertFalse($this->validator->isValid(null));
-        $this->assertArrayHasKey('invalidType', $this->validator->getMessages());
+        $this->assertArrayHasKey(Validator::INVALID_TYPE, $this->validator->getMessages());
     }
 }