+++ test/Validator/TypeBooleanTest.php
@@ -23,15 +23,15 @@ class TypeBooleanTest extends PHPUnit_Framework_TestCase
     public function testInvalidType()
     {
         $this->assertFalse($this->validator->isValid('0'));
-        $this->assertArrayHasKey('invalidType', $this->validator->getMessages());
+        $this->assertArrayHasKey(Validator::INVALID_TYPE, $this->validator->getMessages());
 
         $this->assertFalse($this->validator->isValid(0));
-        $this->assertArrayHasKey('invalidType', $this->validator->getMessages());
+        $this->assertArrayHasKey(Validator::INVALID_TYPE, $this->validator->getMessages());
 
         $this->assertFalse($this->validator->isValid(1));
-        $this->assertArrayHasKey('invalidType', $this->validator->getMessages());
+        $this->assertArrayHasKey(Validator::INVALID_TYPE, $this->validator->getMessages());
 
         $this->assertFalse($this->validator->isValid(null));
-        $this->assertArrayHasKey('invalidType', $this->validator->getMessages());
+        $this->assertArrayHasKey(Validator::INVALID_TYPE, $this->validator->getMessages());
     }
 }